[MNT] `convolution_based` rename (#90)
* kernel to convolution

* kernel to convolution

* wrong move

* init

* docs

* docs2

* notebook

* fix

* missed

---------
