+++ sktime/classification/convolution_based/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+"""Convolution based time series classifiers."""
+
+__all__ = ["RocketClassifier", "Arsenal"]
+
+from sktime.classification.convolution_based._arsenal import Arsenal
+from sktime.classification.convolution_based._rocket_classifier import RocketClassifier
