+++ sktime/classification/convolution_based/_arsenal.py
@@ -101,7 +101,7 @@ class Arsenal(BaseClassifier):
 
     Examples
     --------
-    >>> from sktime.classification.kernel_based import Arsenal
+    >>> from sktime.classification.convolution_based import Arsenal
     >>> from sktime.datasets import load_unit_test
     >>> X_train, y_train = load_unit_test(split="train", return_X_y=True)
     >>> X_test, y_test =load_unit_test(split="test", return_X_y=True)
