+++ sktime/classification/feature_based/__init__.py
@@ -1,9 +1,13 @@
 # -*- coding: utf-8 -*-
-"""Feature based time series classifiers."""
+"""Feature based time series classifiers.
+
+While a bit vague, the contents mostly consist of transformers that extract features
+pipelined to a vector classifier.
+"""
+
 __all__ = [
     "Catch22Classifier",
     "MatrixProfileClassifier",
-    "RandomIntervalClassifier",
     "SignatureClassifier",
     "SummaryClassifier",
     "TSFreshClassifier",
@@ -14,9 +18,6 @@ from sktime.classification.feature_based._catch22_classifier import Catch22Class
 from sktime.classification.feature_based._fresh_prince import FreshPRINCE
 from sktime.classification.feature_based._matrix_profile_classifier import (
     MatrixProfileClassifier,
-)
-from sktime.classification.feature_based._random_interval_classifier import (
-    RandomIntervalClassifier,
 )
 from sktime.classification.feature_based._signature_classifier import (
     SignatureClassifier,
