+++ sktime/classification/kernel_based/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 """Kernel based time series classifiers."""
-__all__ = ["RocketClassifier", "Arsenal", "TimeSeriesSVC"]
 
-from sktime.classification.kernel_based._arsenal import Arsenal
-from sktime.classification.kernel_based._rocket_classifier import RocketClassifier
+__all__ = ["TimeSeriesSVC"]
+
 from sktime.classification.kernel_based._svc import TimeSeriesSVC
