+++ sktime/classification/tests/_classification_test_reproduction.py
@@ -3,6 +3,7 @@ import numpy as np
 from sklearn.ensemble import IsolationForest, RandomForestClassifier
 
 from sktime.classification.compose import ColumnEnsembleClassifier
+from sktime.classification.convolution_based import Arsenal, RocketClassifier
 from sktime.classification.dictionary_based import (
     MUSE,
     WEASEL,
@@ -18,7 +19,6 @@ from sktime.classification.early_classification import (
 from sktime.classification.feature_based import (
     Catch22Classifier,
     MatrixProfileClassifier,
-    RandomIntervalClassifier,
     SignatureClassifier,
     SummaryClassifier,
 )
@@ -26,11 +26,11 @@ from sktime.classification.hybrid import HIVECOTEV1, HIVECOTEV2
 from sktime.classification.interval_based import (
     CanonicalIntervalForest,
     DrCIF,
+    RandomIntervalClassifier,
     RandomIntervalSpectralEnsemble,
     SupervisedTimeSeriesForest,
     TimeSeriesForestClassifier,
 )
-from sktime.classification.kernel_based import Arsenal, RocketClassifier
 from sktime.classification.shapelet_based import ShapeletTransformClassifier
 from sktime.datasets import load_basic_motions, load_unit_test
 from sktime.datatypes._panel._convert import from_nested_to_3d_numpy
