+++ sktime/regression/convolution_based/__init__.py
@@ -2,4 +2,4 @@
 """Kernel based time series regressors."""
 __all__ = ["RocketRegressor"]
 
-from sktime.regression.kernel_based._rocket_regressor import RocketRegressor
+from sktime.regression.convolution_based._rocket_regressor import RocketRegressor
