+++ sktime/regression/convolution_based/_rocket_regressor.py
@@ -1,18 +1,19 @@
 # -*- coding: utf-8 -*-
-"""RandOm Convolutional KErnel Transform (Rocket).
+"""RandOm Convolutional KErnel Transform (Rocket) regressor.
 
-Pipeline classifier using the ROCKET transformer and RidgeClassifierCV estimator.
+Pipeline regressor using the ROCKET transformer and RidgeCV estimator.
 """
 
-__author__ = ["MatthewMiddlehurst", "victordremov", "fkiraly"]
-__all__ = ["RocketClassifier"]
+__author__ = ["fkiraly"]
+__all__ = ["RocketRegressor"]
 
 import numpy as np
-from sklearn.linear_model import RidgeClassifierCV
+from sklearn.linear_model import RidgeCV
 from sklearn.preprocessing import StandardScaler
 
-from sktime.classification._delegate import _DelegatedClassifier
 from sktime.pipeline import make_pipeline
+from sktime.regression._delegate import _DelegatedRegressor
+from sktime.regression.base import BaseRegressor
 from sktime.transformations.panel.rocket import (
     MiniRocket,
     MiniRocketMultivariate,
@@ -22,14 +23,14 @@ from sktime.transformations.panel.rocket import (
 )
 
 
-class RocketClassifier(_DelegatedClassifier):
-    """Classifier wrapped for the Rocket transformer using RidgeClassifierCV.
+class RocketRegressor(_DelegatedRegressor, BaseRegressor):
+    """Regressor wrapped for the Rocket transformer using RidgeCV regressor.
 
-    This classifier simply transforms the input data using the Rocket [1]_
-    transformer and builds a RidgeClassifierCV estimator using the transformed data.
+    This regressor simply transforms the input data using the Rocket [1]_
+    transformer and builds a RidgeCV estimator using the transformed data.
 
     Shorthand for the pipeline
-    `rocket * StandardScaler(with_mean=False) * RidgeClassifierCV(alphas)`
+    `rocket * StandardScaler(with_mean=False) * RidgeCV(alphas)`
     where `alphas = np.logspace(-3, 3, 10)`, and
     where `rocket` depends on params `rocket_transform`, `use_multivariate` as follows:
 
@@ -43,13 +44,13 @@ class RocketClassifier(_DelegatedClassifier):
 
     classes are sktime classes, other parameters are passed on to the rocket class.
 
-    To build other classifiers with rocket transformers, use `make_pipeline` or the
-    pipeline dunder `*`, and different transformers/classifiers in combination.
+    To build other regressors with rocket transformers, use `make_pipeline` or the
+    pipeline dunder `*`, and different transformers/regressors in combination.
 
     Parameters
     ----------
     num_kernels : int, optional, default=10,000
-        The number of kernels for the Rocket transform.
+        The number of kernels the for Rocket transform.
     rocket_transform : str, optional, default="rocket"
         The type of Rocket transformer to use.
         Valid inputs = ["rocket", "minirocket", "multirocket"]
@@ -74,18 +75,12 @@ class RocketClassifier(_DelegatedClassifier):
         The number of classes.
     classes_ : list
         The classes labels.
-    estimator_ : ClassifierPipeline
-        RocketClassifier as a ClassifierPipeline, fitted to data internally
+    estimator_ : RegressorPipeline
+        RocketRegressor as a RegressorPipeline, fitted to data internally
 
     See Also
     --------
-    Rocket
-
-    Notes
-    -----
-    For the Java version, see
-    `TSML <https://github.com/uea-machine-learning/tsml/blob/master/src/main/java/
-    tsml/classifiers/shapelet_based/ROCKETClassifier.java>`_.
+    Rocket, RocketClassifier
 
     References
     ----------
@@ -95,20 +90,19 @@ class RocketClassifier(_DelegatedClassifier):
 
     Examples
     --------
-    >>> from sktime.classification.kernel_based import RocketClassifier
+    >>> from sktime.regression.convolution_based import RocketRegressor
     >>> from sktime.datasets import load_unit_test
     >>> X_train, y_train = load_unit_test(split="train", return_X_y=True)
     >>> X_test, y_test = load_unit_test(split="test", return_X_y=True)
-    >>> clf = RocketClassifier(num_kernels=500)
-    >>> clf.fit(X_train, y_train)
-    RocketClassifier(...)
-    >>> y_pred = clf.predict(X_test)
+    >>> reg = RocketRegressor(num_kernels=500)
+    >>> reg.fit(X_train, y_train)
+    RocketRegressor(...)
+    >>> y_pred = reg.predict(X_test)
     """
 
     _tags = {
         "capability:multivariate": True,
         "capability:multithreading": True,
-        "classifier_type": "kernel",
     }
 
     # valid rocket strings for input validity checking
@@ -134,7 +128,7 @@ class RocketClassifier(_DelegatedClassifier):
         self.n_jobs = n_jobs
         self.random_state = random_state
 
-        super(RocketClassifier, self).__init__()
+        super(RocketRegressor, self).__init__()
 
         if use_multivariate not in self.VALID_MULTIVAR_VALUES:
             raise ValueError(
@@ -172,12 +166,12 @@ class RocketClassifier(_DelegatedClassifier):
         self.multivar_rocket_ = make_pipeline(
             multivar_rocket,
             StandardScaler(with_mean=False),
-            RidgeClassifierCV(alphas=np.logspace(-3, 3, 10)),
+            RidgeCV(alphas=np.logspace(-3, 3, 10)),
         )
         self.univar_rocket_ = make_pipeline(
             univar_rocket,
             StandardScaler(with_mean=False),
-            RidgeClassifierCV(alphas=np.logspace(-3, 3, 10)),
+            RidgeCV(alphas=np.logspace(-3, 3, 10)),
         )
 
         if not use_multivariate:
@@ -210,10 +204,6 @@ class RocketClassifier(_DelegatedClassifier):
         parameter_set : str, default="default"
             Name of the set of test parameters to return, for use in tests. If no
             special parameters are defined for a value, will return `"default"` set.
-            For classifiers, a "default" set of parameters should be provided for
-            general testing, and a "results_comparison" set for comparing against
-            previously recorded results if the general set does not produce suitable
-            probabilities to compare against.
 
         Returns
         -------
@@ -223,7 +213,4 @@ class RocketClassifier(_DelegatedClassifier):
             `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
             `create_test_instance` uses the first (or only) dictionary in `params`.
         """
-        if parameter_set == "results_comparison":
-            return {"num_kernels": 100}
-        else:
-            return {"num_kernels": 20}
+        return {"num_kernels": 20}
